+++ crates/cli-tools/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Minor
 
+- Add `action::AppletMetadata` to print the metadata of an applet
 - Use the toolchain of the release if `rustup` is installed
 - Use the `immediate-abort` panic strategy instead of `build-std-features`
 - Test in `changelog::execute_ci()` that skip counter is zero for releases