+++ crates/cli/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor
 
+- Add `applet-metadata` to print the metadata of an applet
 - Improve help for `host`
 - Do not hardcode `x86_64-unknown-linux-gnu` as the only target