+++ crates/protocol/CHANGELOG.md
@@ -15,6 +15,8 @@
 
 ### Minor
 
+- Add `Api::AppletMetadata` and `applet::Metadata` for the metadata of an applet
+- Bump `Api::AppletInstall` tag to make sure `AppletMetadata` is checked
 - Add `common::Name` for graphic ASCII strings
 - Add `platform::DynInfo` to access the platform info regardless of the device version
 - Add `platform::Info` to also return `common::AppletKind` and platform name